requirements.txt
streamlit
groq
python-dotenv
pandas
matplotlib
seaborn
plotly